Added ramona actor, coins and collection, scaling to tile_test
Added ramona actor to the tile test along with a coin actor, spawned
with the coin objects in the level collectible_layer.  I fixed flipping
of sprites when used in sprite groups, added scaling to sprites and
added collide_rect to allow smaller collision box.

Level was edited to add more coins.